🚚 rename ChatBox to ChatBot
